Conditions v1.2
Changed "temp" to a float value.

+++ conditions.py
@@ -1,5 +1,5 @@
 while True:
-    temp = int(input("Enter the temperature in degrees fahrenheit or enter \"999\" to end the program: "))
+    temp = float(input("Enter the temperature in degrees fahrenheit or enter \"999\" to end the program: "))
 
     if temp == 999:
         break
